fix: replace unnecessary regex replaceAll with literal replace, fix parsing safety
- ManifestGenerator: replaceAll(regex)  replace(literal) for ${applicationId}
- ProjectFilePaths: replaceAll(regex)  replace(literal) for Application class/package substitution
- XmlBuilder: replaceAll(\n)  replace(\n) for newline indent
- BackupFactory: add Matcher.quoteReplacement() to prevent regex replacement issues with special chars in date format
- PropertiesUtil: Integer.parseInt  Float.parseFloat for dimen values that may contain decimals

+++ app/src/main/java/pro/sketchware/utility/PropertiesUtil.java
@@ -44,7 +44,7 @@ public class PropertiesUtil {
             if (TextUtils.isEmpty(value) || reference == null) {
                 return defaultValue;
             }
-            return Integer.parseInt(reference.second);
+            return (int) Float.parseFloat(reference.second);
         } catch (NumberFormatException e) {
             return defaultValue;
         }
